Add a new debug option, --dump-mlds-pred-name.
An MLDS dump contains the C code of an entire module, just as a HLDS dump
by default contains the HLDS of an entire module. For HLDS dumps, we have
long had --dump-hlds-pred-name, which dumped out only the definitions
of the predicates and functions with the given name (or names, if the option
is repeated). This diff adds --dump-mlds-pred-name, which does the same thing
for MLDS dumps.

compiler/options.m:
doc/user_guide.texi:
    Add the new option.

compiler/mlds_to_c.m:
    Add a new predicate to output the definitions of only the named predicates
    and functions.

compiler/mercury_compile_mlds_back_end.m:
    If the new option is given, invoke the new predicate in mlds_to_c.m.