Fix running the tests in a java grade bootcheck.
When bootchecking the test suite is run using the stage 2 compiler, but the
launcher scripts we generate for executables set the CLASSPATH based on the
install path of the compiler.  Add a new environment variable,
MERCURY_STAGE2_LAUNCHER_BASE, that causes the launcher script to be set the
CLASSPATH based on the stage 2 location and directory structure.

Set the new environment variable when bootchecking in the java grade.

compiler/module_cmds.m:
    Handle the new environment variable.

    Add a note about the environment variable not yet being handled on MSYS.
    (I'll fix this separately.)

tools/bootcheck.m:
    Set the new environment variable to the stage 2 directory if bootchecking
    in the java grade.