Fix a bug in --prefer-while-loop-over-jump-mutual.
When we generate tail recursive MLDS code for a set of mutually recursive
procedures, we can wrap up and tie together the bodies of those procedures
using either gotos and labels, or while loops and continues. For C, the
default is the former, but for Java, it is the latter. This diff fixes
a bug in the implementation of the while/continue version. The bug was
that when a procedure succeeded, it executed a break to get out of the
tail recursive loop, but since the break statement was inside a switch,
it broke out of only the switch, and the loop went on, forever.

compiler/ml_proc_gen.m:
    Fix the bug by adding a break after the switch as well. Execution gets
    there only if this is what we want to do.

    Make the predicates for each kind of wrapping up (goto/label and
    while/continue) more self-contained.