Fix the failure of valid/unify_typeinfo_bug in the java grade.
The above test is failing because the runtime does not define a specialised
method pointer interface for arity zero predicates but the code generator
thinks that that it does.  The fix is to change the code generator to use the
generic method pointer interface for zero arity predicates.  (Zero arity
predicates are rare enough that there is no point in having a specialised
interface for them.)

compiler/mlds_to_java.m:
     Make the above change.