Make clause_to_proc a separate pass.
Copying goals from clauses to procedures has to be done after type checking
(which operates on predicates) but before mode checking (which operates on
procedures). Previously, this copying was done within one of the passes
between type and mode checking, to reduce the overall number of passes.
For many years now, it was done within the polymorphism pass.

However, this arrangement has two undesirable consequences.

The first, which was the reason for this change, is that it makes it
much harder to see the effects of the polymorphism pass in HLDS dumps,
since the difference between HLDS dumps before and after the polymorphism
pass contains not just the effects of the polymorphism transformation itself,
but also the population of the goal fields of proc_infos.

The second, which I discovered while doing the change, is that the
polymorphism pass itself is split. It does the vast bulk of its work
before copying clauses to procs, but it does leave some tasks to be done
*after* copying clauses to procs. I don't *know* whether this is
by design or by accident, but I strongly suspect the latter.

compiler/mercury_compile_front_end.m:
    Make copying clauses to procedures its own pass.

    Make the part of polymorphism that runs *after* copying clauses
    to procedures its own pass as well.

compiler/clause_to_proc.m:
    Add a predicate that mercury_compile_front_end can call to do its job
    on every predicate in the module. The setup we use for imported procedures
    is the same as what polymorphism.m used to do for such procedures.

    Make some predicate names more specific.

compiler/polymorphism.m:
    Don't copy clauses to procs in the middle of the polymorphism
    transformation.

    Add a predicate that mercury_compile_front_end.m can call to do the
    post-copying part of the polymorphism transformation.

compiler/introduce_exists_casts.m:
    Add a predicate that polymorphism.m can call to do this module's part of
    the post-copying part of the polymorphism transformation.

compiler/mode_ordering.m:
compiler/modes.m:
compiler/proc_requests.m:
    Conform to the changes above.