Delete if-cross-compiling actions in AC_TRY_RUN calls.
If AC_TRY_RUN cannot proceed because we are cross-compiling, we tended
to silently assume a result, usually the same result as if the
AC_TRY_RUN test had failed. But this might not be correct, or it might
be sub-optimal. It is better for the AC_TRY_RUN to loudly complain so
that the user is aware that something needs to be done.

configure.ac:
    Delete most action-if-cross-compiling arguments in calls to AC_TRY_RUN.