Be more careful with some GCC builtin functions.
Our use of GCC's __builtin_{popcount,clz,ctz} functions in the uint32 module
assumes that sizeof(int) == sizeof(uint32_t).  While this is true on all of the
systems that we target that I am aware of, there's no reason it has to hold in
general.  Modify the macros guarding our use of these functions so that we
only use them if the above condition holds.

configure.ac:
runtime/mercury_conf.h.in:
    Add a new macro that if defined indicates that C's int type is
    exactly 32-bits in size.

library/uint32.m:
    Use the new macro to protect our use of the above builtin functions.