Support editline as an alternative to GNU readline.
Support the use of the editline library as an alternative to GNU readline for
the debugger command-line prompt.  The former has a more permissive license
than the latter.  If licensing is not an issue then GNU readline is the
preferable choice (e.g. feature wise); as such it is the default option.

NOTE: there are (apparently) two editline libraries, one derived from NetBSD
and one from Minix -- this diff adds support for the former.  (Using the latter
should just be a matter of using a different header file and supplying a
different library name to link against, but I haven't tried that yet.)

m4/mercury.m4:
   Add a configuration check for editline.

configure.ac:
    Check the presence of editline.

runtime/mercury_conf.h.in:
    Replace the macro MR_NO_USE_READLINE with MR_USE_READLINE which
    is defined if GNU readline is available and the user has not
    forbidden its use.

    Add a new macro MR_USE_EDITLINE which does the same for editline.

    Add a new macro that is defined if the header editline/readline.h
    is present.

trace/mercury_trace_completion.c:
trace/mercury_trace_readline.c:
    Use editline in place of readline if appropriate.