Fix bug #463.
The compiler has a sanity check that we do not attempt to use a MS C#
compiler in a POSIX (but not Cygwin) environment.  As the MS C# compiler
does now run in POSIX environments this sanity check is out-of-date:
this diff deletes it.

compiler/handle_options.m:
    Delete the out-of-date sanity check.

compiler/compile_target_code.m:
    Conform to the above change.