Strengthen libgrade detection for C# and Java grades.
Runtime detection of installed library grades works by checking for the
presence of a grade specific file in the location where the standard libraries
are installed.  For C and Erlang grades we check for the presence of the
standard library's init file. (This is a good choice since its name and
relative location are consistent across all platforms.)

The C# and Java backends do not generate init files.  Previously, we used the
presence of lib/mercury/modules/csharp directory etc to determine whether those
grades were installed.  As discussed recently on the developers list, this can
lead to bad interactions with package managers if other libraries are also
installed alongside the standard libraries.  This diff changes runtime library
grade detection to check for C# and Java grades by checking for the presence of
the standard library assembly or JAR respectively.  (Unlike in C grades, these
will have consistent names across platforms.)

compiler/mercury_compile_main.m:
     Detect installed C# and Java grades by looking for the standard library
     assembly or JAR respectively.