Implement abstract insts and modes for use in .int3 files.
The syntax is

    :- abstract_inst((inst_name(instvar1, ..., instvarN))).
    :- abstract_mode((inst_name(instvar1, ..., instvarN))).

The rationale for that syntax is as follows.

First, I ruled out using the same scheme as for types, which would yield
the syntax

    :- inst (inst_name(V_1, ..., V_N)).
    :- mode (inst_name(V_1, ..., V_N)).

It would work for insts, but not for modes, because ":- mode p." would be
ambiguous: one couldn't tell whether it was defining a *mode* named p
of arity 0, or it was declaring a mode for a *predicate* named p
of arity 0. And I thought it made more sense for the syntax of insts
to follow the syntax of modes than to follow the syntax of types.
Hence the addition of the "abstract_" prefix. Since I intend to put
abstract insts and modes automatically into .int3 files, but NOT to add them
to the user visible part of the language, this should be fine.

Second, I wanted to avoid changing the set of operators in the language,
which is why there is no space after "abstract_inst" and "abstract_mode".

Third, there is a redundant set of parentheses around the argument
of those function symbols, just as there is around the inst name in
the non-abstract inst and mode definitions in .int3 file right now,
for the same reason: to guard against case that the inst or mode name
*is* an operator.

Fourth, the reason for listing the argument inst vars, instead of adding
a "/arity" suffix to the name, is to reuse the existing machinery
for reading in and writing out the heads of inst and mode definitions.

NOTE: the compiler still puts inst and mode definitions into .int3 files
in their original, non-abstract state. Changing that is for a later diff.

compiler/prog_item.m:
    Change the representation of inst and mode definitions to allow
    the inst or mode being defined to be abstract. These abstract versions
    do not make it into the HLDS.

compiler/prog_data.m:
compiler/hlds_inst_mode.m:
    We used to have a representation for abstract inst definitions inside
    both the parse tree and the HLDS, but any attempt to add the parse tree
    version to the HLDS would yield a compiler abort.

    Delete these representations, since they never worked, and are now not
    needed.

compiler/parse_inst_mode_defn.m:
compiler/parse_item.m:
    Parse the new syntax for abstract inst and mode definitions.

    Stop parsing the old syntax, which we have never used.

compiler/parse_tree_out.m:

compiler/add_mode.m:
    Add only nonabstract insts and modes to the HLDS. The abstract versions
    are used only during module qualification, which is done on the parse tree.

compiler/add_mutable_aux_preds.m:
compiler/equiv_type.m:
compiler/hlds_out_module.m:
compiler/inst_check.m:
compiler/inst_user.m:
compiler/intermod.m:
compiler/mode_util.m:
compiler/module_qual.qualify_items.m:
compiler/recompilation.usage.m:
compiler/unused_imports.m:
    Conform to the changes above.