Make some error messages order-independent.
The old code in modules.m that was recently delete preserved the order
of the items in the source code. The new code that replaced it groups
all the items in a section together. If the source code follows any
pattern other than "one interface section, one implementation section",
then this can change the relative order of two items.

Previously we assumed that items are added to the HLDS in order of their
occurrence in the source code. This diff eliminates that assumption
when we generate error messages that refer to two or more items.

compiler/add_class.m:
    Eliminate this adding-items-by-source-context-order assumption
    when diagnosing duplicate instance declarations.

compiler/add_type.m:
    Eliminate this adding-items-by-source-context-order assumption
    when diagnosing duplicate type declarations and definitions
    (both Mercury types and foreign types).

compiler/make_hlds_error.m:
    Eliminate this adding-items-by-source-context-order assumption
    when diagnosing duplicate declarations of many different kinds
    of items.

compiler/style_checks.m:
    Eliminate this adding-items-by-source-context-order assumption
    when diagnosing non-adjacent argument type and mode declarations
    for the same predicate or function.

compiler/termination.m:
    Eliminate this adding-items-by-source-context-order assumption
    when diagnosing inconsistent termination assertions
    for mutually recursive procedures. Simplify the code
    that generates the error messages doing this.

tests/invalid/bug476.err_exp:
    Expect an error message that uses the phrase
    "previous definition" correctly.