Add 64-bit versions of seek and offset on binary streams.
The current versions of the seek and offset operations on binary streams use an
int to represent the offset.  On systems where Mercury's int type is 32-bit
this effectively limits the sizes of file we can handle with these operations
to approx. 2GB.  Additionally, the underlying operations are broken on 64-bit
Windows because fseek() and ftell() on Windows only use 32-bit values.  (We
should be using _fseeki64() and _ftelli64() on Windows -- that will be a
separate change however.)

This diff introduces versions of the seek and offset operations that use the
int64 type to represent the stream offset.  It also changes the underlying
infrastructure (e.g. the Java implementation of Mercury binary file streams) to
represent offsets using only a 64-bit type.  The existing stream and offset
operations are now implemented by _casting_ the int64 offsets to ints -- this
is wrong on some platforms, but it is the existing behaviour.

library.io.m:
     Add the predicates seek64_binary_{input,output}/5 and
     binary_{input,output}_stream_offset64/4.

     Shift the input stream versions of the above operations into the correct
     section; for some reason they were listed with the predicates that
     operate on binary *output* streams.

     Add XXXs about related issues that need to fixed.

     Use 64-bit integers only for seek and offset operations on binary
     streams.