Convert the NEWS file to markdown and begin restructuring it.
This diff makes two changes to the NEWS file. The first is to convert it into
markdown. There are a number of reasons we want to do this:

   1. It allows us to automate the conversion to HTML.
   2. It will look nicer on github.
   3. It will still look (reasonably) nice as plain text.

The second change is to begin restructuring the contents of the file along the
lines proposed in the thread on the developers list titled
"Renaming the NEWS file -> NEWS.md" from February 2018. Specifically, this
change implements the proposals that the be exactly one entry in the file for
any library module and that those entries be sorted by module name.

NEWS:
   Make the above changes.