Fix the clang version detection macro.
We currently determine which version of clang we are using by compiling a small
C program and examining the value of the macros __clang_major__ etc.  There are
two problems with this.

1. As noted recently on the developers list, having to run the program
interferes with cross-compilation.

2. Apple vendored clangs "lie" about their version.  For the Apple versions the
macros __clang__major__ etc, seem to be set to the XCode version, not the
underlying clang version.

Fix both the above issues by parsing the output of -dumpversion as we do with
GCC.  (The Apple versions do report the actual clang version when using that
flag.)

m4/mercury.m4:
   Make the above change.