Make M.{c,java,etc} depend on M.int.
To enable argument packing, we need to ensure that all modules that use
a type (say T) agree on its representation. The design I am working to
implement designates the .int file of the module that defines the type
(call it M) as the single source of truth for this. Every module that
imports the definition of the type, even through intermodule optimization,
gets that definition from M.int. The only module that depends on the type
that does not get its definition from M.int is M itself.

Before I can add code that makes the compiler get the representation of
T from M.int even when generating for M itself, I have to make sure that
at the time of that compiler invocation, M.int will (a) exist, and (b)
be up-to-date. That is why this change makes M.int a prerequisite
for generating target language code for M both when using mmc --make
and when using mmake.

compiler/make.dependencies.m:
    Add M.int to the list of dependencies for M's code generation.

compiler/write_deps_file.m:
    Add M.int as an extra source for the mmake rule containing
    code generation dependencies.

    Add an XXX.

tests/invalid/Mercury.options:
    Delete --compile-only from a test case's options. Since we now
    generate an interface file for the test case, the --compile-only
    would otherwise clash with the --make-interface option needed for that.