Start transition to unnamed type_infos in foreign_procs.
Foreign_procs that want to access the type_infos of their arguments
could always do so using compiler-defined variables with names such as
TypeInfo_for_T, where T is the name of the relevant type variable.
This scheme is simple and easy to use, but it prevents us from
erasing the names of type variables from interface files.

Start a transition to a state where foreign_procs can refer to
type_infos via names such as TypeInfo_In_1, TypeInfo_In_2 etc
for type_infos that are input to the foreign_proc, and names such as
TypeInfo_Out_1, TypeInfo_Out_2 etc for type_infos that are output from it
(for the very rare case of a foreign_proc for an existentially typed
predicate or function). In both cases, the numbers are given in the
same order as the type_info arguments themselves, as added to the
front of the visible arg list by polymorphism. Within both the set of
input type_info args and the set of output type_info args, this order
in turn is given by the depth first left to right traversal of the predicate
or function's type signature. The scheme uses separate numbering for
inputs and outputs because this does not hold when comparing an input
type_info with an output type_info, due to the syntax for existential types.

compiler/ml_foreign_proc_gen.m:
compiler/pragma_c_gen.m:
    Implement the initial step of the transition. This stage declares
    *both* names for each type_info, e.g. both TypeInfo_for_T and
    TypeInfo_In_1 or TypeInfo_Out_1.

    For inputs, it assigns TypeInfo_for_T, which is set as normal,
    to TypeInfo_In_1 before the user code is entered. This should allow us
    to modify code that uses TypeInfo_for_T to switch to TypeInfo_In_1
    at a time of our choosing.

    For outputs, if the code does not refer to TypeInfo_Out_1, it assigns
    TypeInfo_for_T to TypeInfo_Out_1 after the user code has exited.
    This assignment initially will be redundant, but its presence
    simplifies the correctness argument. If the code *does* refer to
    TypeInfo_Out_1, which it will after we have updated the foreign_proc
    in question, we assign it to TypeInfo_for_T instead, which is
    where the code generator will pick its value from, as normal.

    After we have updated all foreign_procs that refer to type_infos,
    we should be able to remove all references to variable names that
    follow the old naming scheme from the generated code.

    Simplify some of the related code.

compiler/options.m:
    Add a synonym for compiler_sufficiently_recent that allows configure
    to detect installed compilers that contain this change.