Output INT32_MIN instead of INT32_C(-2147483648).
When -2147483648i32 is used in a Mercury source file,
the generated C source file would contain:

    INT32_C(-2147483648)

When compiled with gcc in C89/C90 mode (the default on older versions)
on a 32-bit platform, gcc would generate this warning:

    warning: this decimal constant is unsigned only in ISO C90

because the unsuffixed decimal constant 2147483648 is given the type
`unsigned long int' according to C89/C90 rules.
We previously saw the same warning with Mercury `int' literals.

compiler/c_util.m:
    In C, write out the value of the min_int32 as the symbolic
    constant INT32_MIN. This mirrors behaviour for min_int64.