Add new option --show-definition-extents.
When specified, the compiler will generate an overview of the sequence
of predicates and functions in the module being compiled, and put it
in a file with the suffix ".defn_extents". For tests/benchmarks/nrev,
this file looks like this:

    main/2              26 to     26
    main1/1             28 to     30
    main3/3             35 to     37
    data/1              42 to     42
    nreverse/2          48 to     50
    concatenate/3       55 to     57
    print_list/3        62 to     64
    print_list_2/3      76 to     80

This overview can help spot situations in which related predicates
and functions are not grouped together.

compiler/options.m:
    Add the new option.

compiler/mercury_compile_main.m:
    If the new option is specified, invoke hlds_defns.m to do the task
    that it asks for.

compiler/hlds_defns.m:
    Implement the new option, mostly by slightly generalizing the old code
    that implements the related option --show-definition-line-counts,
    and wrapping it up in a small amount of new code.

doc/user_guide.texi:
NEWS:
    Document and announce the new option.