Allow reporting of stats to a stream for C.
Until now, all the predicates that reported statistics wrote their output
to stderr; users could not choose another stream. This diff adds ways
to report those same statistics to a stream chosen by the user, though
for now, only when targeting C.

The code for reporting statistics used to be in benchmarking.m, in the
form of a foreign_decl/foreign_code pair that declared and defined the
needed data structures and functions, and foreign_procs that called them.
All this code had stderr hard-coded as the destination stream.

As long as the output is done by C code (as opposed to the C code
constructing a big string containing all the stuff to be output),
directing the output to a user-supplied io.text_output_stream requires
converting that text_output_stream to a value of the C "FILE *" type.
This requires access to the actual definition of the io.text_output_stream
type, which is private to io.m. We could export it to benchmarking.m,
but that is undesirable. Instead, this diff moves the code that used
to be in benchmarking.m's foreign_decl/foreign_code pair to a new
module in the runtime, mercury_report_stats.[ch], which allows
foreign_procs in both benchmarking.m and io.m to use it.

We used to have

    impure report_stats/0       in benchmarking.m
    report_stats/2              in io.m
    report_stats/3              in io.m

The difference between the two versions in io.m is a "selector" argument
that specifies which kind of stats to print. Adding a text_output_stream
argument to report_stats/2 would create report_stats/3, which clashes
with this existing name/arity pair. This diff therefore renames
io.report_stats/2 as io.report_standard_stats.

library/benchmarking.m:
    Delete the code moved to mercury_report_stats.[ch]. Call
    ML_report_full_memory_stats even if MR_MPROF_PROFILE_MEMORY is not defined
    (see next entry).

runtime/mercury_report_stats.[ch]:
    New module containing the code moved from benchmarking.m, with
    ML_report_stats renamed to ML_report_standard_stats, and then
    all ML_ prefixes replaced with MR_ prefixes.

    There is one substantive change. In benchmarking.m, the
    ML_report_full_memory_stats function was always declared, but it was
    defined, and used, only inside "#ifdef MR_MPROF_PROFILE_MEMORY".
    Change this so that it is always defined and used, but prints
    a message about memory profiling not being available if
    MR_MPROF_PROFILE_MEMORY is not defined. This is preferable to the empty
    output that calls to report_full_memory_stats used to generate in such
    situations.

runtime/Mmakefile:
    Add the new header and source files to the lists of header and source
    files respectively.

library/io.m:
    Make the changes described above.

    Provide specific predicates for the full_memory and tabling stats,
    as well as for the standard stats. For now, these are implemented
    only for C.

    Fix inconsistent indentation.

NEWS:
    Announce the rename of io.report_stats/2 to io.report_standard_stats/2,
    but do not announce the other new predicates in io.m yet. That should
    happen when they work for C# and Java as well.

compiler/file_util.m:
    Conform to the change from report_stats to report_standard_stats.

 	runtime/.r1