Fix --track-flags.
Commit 181ada0dbf2899b388ae9155083a3050dd027cc1 changed many options
from normal options to special options. This revealed a bug in the
implementation of --track-flags. We previously computed the hash of a
list of (option, option_data) pairs to detect if options differ from a
previous run. However, for special options, the option_data is always
bool_special, int_special, etc. so whatever value those options
were set to had no effect on the hash value.

compiler/make.m:
    Include the opt_tuple table when computing the hash for
    `.track_flags' files.