Ensure make does not try to run cross-compiled binaries.
When cross-compiling, `make' must not find cross-compiled binaries in
the workspace and try to execute them. For example, the problem arises
when cross-compiling for Linux/aarch64 from Linux/x86-64. `make' would
find e.g. util/mkinit (for an incompatible platform) and try to run it.

The problem did not arise in the special (but common) case of
cross-compiling for Windows on a Linux host: the Windows binaries would
be given a .exe extension, and `make' on Linux would not try to run
<foo>.exe anyway.

Mmakefile:
    Exclude util and mfilterjavac from the PATH when cross-compiling.