Associate a type with an inst.
Many predicates working on the inst of a variable (or part of the value of
a variable) need to know the type of (that part of) the variable, but
in some contexts, the type is not available at compile time. (One instance
is the type of an existentially typed argument.) Because of this, the
compiler used to pass a maybe(mer_type) with each mer_inst.

Act on an old XXX to change this to pass a mer_type with each mer_inst,
with the mer_type being a type variable if the type is not known statically.
This is conceptually cleaner, and should also be faster, due to avoiding
yes() wrappers in the usual case where the type *is* available.

compiler/inst_match.m:
compiler/inst_merge.m:
compiler/inst_test.m:
    As above.

compiler/inst_util.m:
    Update the names as well as the types of two utility predicates.

    Add a function for returning the canonical representation of no type
    information being available.

compiler/instmap.m:
    Conform to the changes above.

compiler/type_util.m:
    Add a new utility predicate for use by inst_match.m. The old code
    in inst_match.m called a nondet predicate as a semidet test
    by throwing away all its output arguments. The new utility predicate
    avoids the wasted work of constructing those arguments.